% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{gb_build_corpus}
\alias{gb_build_corpus}
\title{Build corpus/tif.}
\usage{
gb_build_corpus(x, as_fulltext = F)
}
\arguments{
\item{x}{A character vvector}

\item{as_fulltext}{boolean}
}
\value{
A data frame
}
\description{
Build corpus/tif.
}
