% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gb_search}
\alias{gb_search}
\title{Search gutenberg}
\usage{
gb_search(x, pattern, ignore.case = T, cores = 5)
}
\arguments{
\item{x}{A character vector}

\item{pattern}{a character string}

\item{cores}{an integer}
}
\value{
A data frame
}
\description{
Search gutenberg
}
