% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{gb_subset}
\alias{gb_subset}
\title{Filter gutenberg corpus by meta}
\usage{
gb_subset(
  x = NULL,
  min_birth_year = NULL,
  max_birth_year = NULL,
  subj = NULL,
  lang = "en",
  filepath = "/home/jtimm/gutenberg/data/sentences/"
)
}
\arguments{
\item{x}{NULL, or integer}

\item{min_birth_year}{NULL, or integer}

\item{max_birth_year}{max_birth_year}

\item{subj}{NULL, or char string}

\item{lang}{NULL, or char string}

\item{filepath}{A character string}
}
\value{
A character vector
}
\description{
Filter gutenberg corpus by meta
}
