% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gbsearch}
\alias{gbsearch}
\title{Search gutenberg}
\usage{
gbsearch(
  pattern = "^I believe",
  x = "/home/jtimm/gutenberg/data/sentences/",
  book_sample = 25000,
  ignore.case = T,
  cores = 12
)
}
\arguments{
\item{pattern}{a character string}

\item{x}{a file path}

\item{book_sample}{An integer}

\item{cores}{an integer}
}
\value{
A data frame
}
\description{
Search gutenberg
}
