% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickly.R
\name{quickly}
\alias{quickly}
\title{Search gutenberg quickly}
\usage{
quickly(
  pattern,
  x = 10000,
  ignore.case = T,
  cores = 6,
  min_birth_year = NULL,
  max_birth_year = NULL,
  subj = NULL
)
}
\arguments{
\item{pattern}{a character string}

\item{x}{An integer}

\item{ignore.case}{a Boolean}

\item{cores}{an integer}
}
\value{
A data frame
}
\description{
Search gutenberg quickly
}
